chore: remove http01 from http util
