+++ src/sinks/http.rs
@@ -3,7 +3,7 @@ use crate::{
     event::{self, Event},
     sinks::util::{
         encoding::{EncodingConfig, EncodingConfiguration},
-        http2::{Auth, BatchedHttpSink, HttpClient, HttpSink},
+        http::{Auth, BatchedHttpSink, HttpClient, HttpSink},
         service2::TowerRequestConfig,
         BatchBytesConfig, Buffer, Compression, UriSerde,
     },
@@ -290,7 +290,7 @@ mod tests {
     use crate::{
         assert_downcast_matches,
         sinks::http::HttpSinkConfig,
-        sinks::util::http2::HttpSink,
+        sinks::util::http::HttpSink,
         sinks::util::test::build_test_server,
         test_util::{next_addr, random_lines_with_stream, runtime, shutdown_on_idle},
         topology::config::SinkContext,
