+++ src/sinks/util/http.rs
@@ -1,6 +1,6 @@
 use super::{
-    retries::{RetryAction, RetryLogic},
-    service::{TowerBatchedSink, TowerRequestSettings},
+    retries2::{RetryAction, RetryLogic},
+    service2::{TowerBatchedSink, TowerRequestSettings},
     sink, Batch, BatchSettings,
 };
 use crate::{
@@ -9,32 +9,35 @@ use crate::{
     tls::{tls_connector_builder, MaybeTlsSettings},
     topology::config::SinkContext,
 };
-use bytes::Bytes;
-use futures::compat::Future01CompatExt;
-use futures01::{Async, AsyncSink, Future, Poll, Sink, StartSend, Stream};
-use http::header::HeaderValue;
-use http::{Request, StatusCode};
-use hyper::body::{Body, Payload};
-use hyper::client::HttpConnector;
-use hyper::Client;
-use hyper_openssl::HttpsConnector;
+use bytes05::{Buf, Bytes};
+use futures::future::BoxFuture;
+use futures01::{Async, AsyncSink, Poll as Poll01, Sink, StartSend};
+use http02::header::HeaderValue;
+use http02::{Request, StatusCode};
+use hyper13::body::{self, Body, HttpBody};
+use hyper13::client::HttpConnector;
+use hyper13::Client;
+use hyper_openssl08::HttpsConnector;
 use serde::{Deserialize, Serialize};
-use std::{fmt, sync::Arc};
-use tokio01::executor::DefaultExecutor;
-use tower::Service;
+use std::{
+    fmt,
+    sync::Arc,
+    task::{Context, Poll},
+};
+use tower03::Service;
 use tracing::Span;
 use tracing_futures::Instrument;
 
-pub type Response = http::Response<Bytes>;
-pub type Error = hyper::Error;
-pub type HttpClientFuture = <HttpClient as Service<http::Request<Body>>>::Future;
+pub type Response = http02::Response<Bytes>;
+pub type Error = hyper13::Error;
+pub type HttpClientFuture = <HttpClient as Service<http02::Request<Body>>>::Future;
 
 pub trait HttpSink: Send + Sync + 'static {
     type Input;
     type Output;
 
     fn encode_event(&self, event: Event) -> Option<Self::Input>;
-    fn build_request(&self, events: Self::Output) -> http::Request<Vec<u8>>;
+    fn build_request(&self, events: Self::Output) -> http02::Request<Vec<u8>>;
 }
 
 /// Provides a simple wrapper around internal tower and
@@ -54,7 +57,7 @@ pub struct BatchedHttpSink<T, B, L = HttpRetryLogic>
 where
     B: Batch,
     B::Output: Clone + Send + 'static,
-    L: RetryLogic<Response = hyper::Response<Bytes>> + Send + 'static,
+    L: RetryLogic<Response = http02::Response<Bytes>> + Send + 'static,
 {
     sink: Arc<T>,
     inner: TowerBatchedSink<HttpBatchService<B::Output>, B, L, B::Output>,
@@ -94,7 +97,7 @@ impl<T, B, L> BatchedHttpSink<T, B, L>
 where
     B: Batch,
     B::Output: Clone + Send + 'static,
-    L: RetryLogic<Response = hyper::Response<Bytes>, Error = hyper::Error> + Send + 'static,
+    L: RetryLogic<Response = http02::Response<Bytes>, Error = hyper13::Error> + Send + 'static,
     T: HttpSink<Input = B::Input, Output = B::Output>,
 {
     pub fn with_retry_logic(
@@ -126,7 +129,7 @@ where
     B: Batch,
     B::Output: Clone + Send + 'static,
     T: HttpSink<Input = B::Input, Output = B::Output>,
-    L: RetryLogic<Response = hyper::Response<Bytes>> + Send + 'static,
+    L: RetryLogic<Response = http02::Response<Bytes>> + Send + 'static,
 {
     type SinkItem = crate::Event;
     type SinkError = crate::Error;
@@ -147,7 +150,7 @@ where
         Ok(AsyncSink::Ready)
     }
 
-    fn poll_complete(&mut self) -> Poll<(), Self::SinkError> {
+    fn poll_complete(&mut self) -> Poll01<(), Self::SinkError> {
         if let Some(item) = self.slot.take() {
             if let AsyncSink::NotReady(item) = self.inner.start_send(item)? {
                 self.slot = Some(item);
@@ -167,8 +170,9 @@ pub struct HttpClient<B = Body> {
 
 impl<B> HttpClient<B>
 where
-    B: Payload + Send + 'static,
+    B: HttpBody + Send + 'static,
     B::Data: Send,
+    B::Error: Into<crate::Error>,
 {
     pub fn new(
         resolver: Resolver,
@@ -190,9 +194,7 @@ where
             Ok(())
         });
 
-        let client = hyper::Client::builder()
-            .executor(DefaultExecutor::current())
-            .build(https);
+        let client = Client::builder().build(https);
 
         let version = crate::get_version();
         let user_agent = HeaderValue::from_str(&format!("Vector/{}", version))
@@ -207,22 +209,23 @@ where
         })
     }
 
-    pub async fn send(&mut self, request: Request<B>) -> crate::Result<http::Response<Body>> {
-        self.call(request).compat().await.map_err(Into::into)
+    pub async fn send(&mut self, request: Request<B>) -> crate::Result<http02::Response<Body>> {
+        self.call(request).await.map_err(Into::into)
     }
 }
 
 impl<B> Service<Request<B>> for HttpClient<B>
 where
-    B: Payload + Send + 'static,
+    B: HttpBody + Send + 'static,
     B::Data: Send,
+    B::Error: Into<crate::Error>,
 {
-    type Response = http::Response<Body>;
+    type Response = http02::Response<Body>;
     type Error = Error;
-    type Future = Box<dyn Future<Item = Self::Response, Error = Self::Error> + Send + 'static>;
+    type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
 
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        Ok(().into())
+    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Ok(()).into()
     }
 
     fn call(&mut self, mut request: Request<B>) -> Self::Future {
@@ -236,19 +239,20 @@ where
 
         debug!(message = "sending request.", uri = %request.uri(), method = %request.method());
 
-        let fut = self
-            .client
-            .request(request)
-            .inspect(|res| {
-                debug!(
+        let response = self.client.request(request);
+
+        let fut = async move {
+            let res = response.await?;
+            debug!(
                     message = "response.",
                     status = ?res.status(),
                     version = ?res.version(),
-                )
-            })
-            .instrument(self.span.clone());
+            );
+            Ok(res)
+        }
+        .instrument(self.span.clone());
 
-        Box::new(fut)
+        Box::pin(fut)
     }
 }
 
@@ -274,14 +278,14 @@ impl<B> fmt::Debug for HttpClient<B> {
 #[derive(Clone)]
 pub struct HttpBatchService<B = Vec<u8>> {
     inner: HttpClient<Body>,
-    request_builder: Arc<dyn Fn(B) -> hyper::Request<Vec<u8>> + Sync + Send>,
+    request_builder: Arc<dyn Fn(B) -> hyper13::Request<Vec<u8>> + Sync + Send>,
 }
 
 impl<B> HttpBatchService<B> {
     pub fn new(
         resolver: Resolver,
         tls_settings: impl Into<MaybeTlsSettings>,
-        request_builder: impl Fn(B) -> hyper::Request<Vec<u8>> + Sync + Send + 'static,
+        request_builder: impl Fn(B) -> hyper13::Request<Vec<u8>> + Sync + Send + 'static,
     ) -> HttpBatchService<B> {
         let inner =
             HttpClient::new(resolver, tls_settings).expect("Unable to initialize http client");
@@ -296,36 +300,38 @@ impl<B> HttpBatchService<B> {
 impl<B> Service<B> for HttpBatchService<B> {
     type Response = Response;
     type Error = Error;
-    type Future = Box<dyn Future<Item = Self::Response, Error = Self::Error> + Send + 'static>;
+    type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
 
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        Ok(().into())
+    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Ok(()).into()
     }
 
     fn call(&mut self, body: B) -> Self::Future {
         let request = (self.request_builder)(body).map(Body::from);
-        let fut = self.inner.call(request).and_then(|r| {
-            let (parts, body) = r.into_parts();
-            body.concat2()
-                .map(|b| hyper::Response::from_parts(parts, b.into_bytes()))
-        });
 
-        Box::new(fut)
+        let response = self.inner.call(request);
+        let fut = async move {
+            let res = response.await?;
+            let (parts, body) = res.into_parts();
+            let mut body = body::aggregate(body).await?;
+            Ok(hyper13::Response::from_parts(parts, body.to_bytes()))
+        };
+
+        Box::pin(fut)
     }
 }
 
-impl<T: fmt::Debug> sink::Response for http::Response<T> {
+impl<T: fmt::Debug> sink::Response for http02::Response<T> {
     fn is_successful(&self) -> bool {
         self.status().is_success()
     }
 }
-
 #[derive(Clone)]
 pub struct HttpRetryLogic;
 
 impl RetryLogic for HttpRetryLogic {
-    type Error = hyper::Error;
-    type Response = hyper::Response<Bytes>;
+    type Error = hyper13::Error;
+    type Response = hyper13::Response<Bytes>;
 
     fn is_retriable_error(&self, error: &Self::Error) -> bool {
         error.is_connect() || error.is_closed()
@@ -357,14 +363,14 @@ pub enum Auth {
 
 impl Auth {
     pub fn apply<B>(&self, req: &mut Request<B>) {
-        use headers::HeaderMapExt;
+        use headers03::HeaderMapExt;
 
         match &self {
             Auth::Basic { user, password } => {
-                let auth = headers::Authorization::basic(&user, &password);
+                let auth = headers03::Authorization::basic(&user, &password);
                 req.headers_mut().typed_insert(auth);
             }
-            Auth::Bearer { token } => match headers::Authorization::bearer(&token) {
+            Auth::Bearer { token } => match headers03::Authorization::bearer(&token) {
                 Ok(auth) => req.headers_mut().typed_insert(auth),
                 Err(error) => error!(message = "invalid bearer token", %token, %error),
             },
@@ -376,11 +382,12 @@ impl Auth {
 mod test {
     use super::*;
     use crate::test_util::runtime;
-    use futures01::{Future, Sink, Stream};
-    use http::Method;
-    use hyper::service::service_fn;
-    use hyper::{Body, Response, Server, Uri};
-    use tower::Service;
+    use bytes05::Buf;
+    use futures01::{Future, Stream};
+    use http02::Method;
+    use hyper13::service::{make_service_fn, service_fn};
+    use hyper13::{Body, Response, Server, Uri};
+    use tower03::Service;
 
     #[test]
     fn util_http_retry_logic() {
@@ -413,47 +420,53 @@ mod test {
 
         let request = b"hello".to_vec();
         let mut service = HttpBatchService::new(resolver, None, move |body: Vec<u8>| {
-            let mut builder = hyper::Request::builder();
-            builder.method(Method::POST);
-            builder.uri(uri.clone());
-            builder.body(body.into()).unwrap()
+            Request::builder()
+                .method(Method::POST)
+                .uri(uri.clone())
+                .body(body.into())
+                .unwrap()
         });
-
-        let req = service.call(request);
 
         let (tx, rx) = futures01::sync::mpsc::channel(10);
 
-        let new_service = move || {
+        let new_service = make_service_fn(move |_| {
             let tx = tx.clone();
 
-            service_fn(move |req: hyper::Request<Body>| -> Box<dyn Future<Item = Response<Body>, Error = String> + Send> {
-                let tx = tx.clone();
+            let svc = service_fn(move |req| {
+                let mut tx = tx.clone();
 
-                Box::new(req.into_body().map_err(|_| "".to_string()).fold::<_, _, Result<_, String>>(vec![], |mut acc, chunk| {
-                    acc.extend_from_slice(&chunk);
-                    Ok(acc)
-                }).and_then(move |v| {
-                    let string = String::from_utf8(v).map_err(|_| "Wasn't UTF-8".to_string());
-                    tx.send(string).map_err(|_| "Send error".to_string())
-                }).and_then(|_| {
-                    futures01::future::ok(Response::new(Body::from("")))
-                }))
-            })
-        };
+                async move {
+                    let body = hyper13::body::aggregate(req.into_body())
+                        .await
+                        .map_err(|e| format!("error: {}", e))?;
+                    let string = String::from_utf8(body.bytes().into())
+                        .map_err(|_| "Wasn't UTF-8".to_string())?;
+                    tx.try_send(string).map_err(|_| "Send error".to_string())?;
 
-        let server = Server::bind(&addr)
-            .serve(new_service)
-            .map_err(|e| eprintln!("server error: {}", e));
+                    Ok::<_, crate::Error>(Response::new(Body::from("")))
+                }
+            });
+
+            async move { Ok::<_, std::convert::Infallible>(svc) }
+        });
 
         let mut rt = runtime();
 
-        rt.spawn(server);
+        rt.spawn_std(async move {
+            if let Err(e) = Server::bind(&addr).serve(new_service).await {
+                eprintln!("server error: {}", e);
+            }
+        });
 
-        rt.block_on(req).unwrap();
+        rt.block_on_std(async move {
+            tokio::time::delay_for(std::time::Duration::from_millis(50)).await;
+            service.call(request).await
+        })
+        .unwrap();
 
         let _ = rt.shutdown_now();
 
         let (body, _rest) = rx.into_future().wait().unwrap();
-        assert_eq!(body.unwrap().unwrap(), "hello");
+        assert_eq!(body.unwrap(), "hello");
     }
 }
