+++ src/transforms/aws_ec2_metadata.rs
@@ -2,7 +2,7 @@ use super::Transform;
 use crate::{
     event::Event,
     hyper::body_to_bytes,
-    sinks::util::http2::HttpClient,
+    sinks::util::http::HttpClient,
     topology::config::{DataType, TransformConfig, TransformContext, TransformDescription},
 };
 use bytes::Bytes;
